fix: calculator type errors + split CI into web/apk workflows
Type fixes:
- effectiveInvRate: .toDouble() on math.pow result
- diff.abs().toDouble() for double parameter
- .clamp(0.0, 99.0) for double clamp bounds

CI/CD split:
- web.yml: auto-runs on every push, deploys to GitHub Pages
- apk.yml: MANUAL only via workflow_dispatch button
- Removed old build.yml that ran both together

+++ zerotab-app/lib/features/calculator/screens/rupee_decision_screen.dart
@@ -183,7 +183,7 @@ _DecisionResult runDecision({
       : grossGain * 0.15; // STCG 15% if < 1 year
 
   final fvAfterTax = fvLumpsum - ltcgTax;
-  final effectiveInvRate = (math.pow(fvAfterTax / extraAmount, 12.0 / remainingMonths) - 1) * 12 * 100;
+  final effectiveInvRate = ((math.pow(fvAfterTax / extraAmount, 12.0 / remainingMonths) - 1) * 12 * 100).toDouble();
 
   // Invest benefit = after-tax future value minus original amount, discounted to PV
   final investBenefit = fvAfterTax - extraAmount;
@@ -196,7 +196,7 @@ _DecisionResult runDecision({
       : diff > 0 ? _Decision.invest : _Decision.prepay;
 
   final reasoning = _buildReasoning(
-    decision, effectiveLoanRate, effectiveInvRate, diff.abs(),
+    decision, effectiveLoanRate, effectiveInvRate, diff.abs().toDouble(),
     monthsSaved, isHomeLoan, isOldTaxRegime, taxBracket,
   );
 
@@ -204,9 +204,9 @@ _DecisionResult runDecision({
     decision:          decision,
     prepayBenefit:     prepayBenefit,
     investBenefit:     investBenefit,
-    netDifference:     diff.abs(),
+    netDifference:     diff.abs().toDouble(),
     effectiveLoanRate: effectiveLoanRate,
-    effectiveInvRate:  effectiveInvRate.clamp(0, 99),
+    effectiveInvRate:  effectiveInvRate.clamp(0.0, 99.0),
     monthsSaved:       monthsSaved,
     futureValue:       fvAfterTax,
     interestSaved:     interestSaved,
